requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1